#165553347 Credit bank account endpoint

+++ server/db/queries.js
@@ -29,7 +29,7 @@ const createQueries = {
   transactionTable: `CREATE TABLE IF NOT EXISTS
     transactions(
       id SERIAL PRIMARY KEY,
-      account_number INTEGER NOT NULL,
+      account_number BIGINT NOT NULL,
       transaction_date TIMESTAMP,
       amount NUMERIC(15, 2) NOT NULL,
       cashier INTEGER NOT NULL,
